12345_updated_code

+++ src/main/java/com/example/repository/MessageRepository.java
@@ -6,6 +6,6 @@ import com.example.entity.Message;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-    public interface MessageRepository extends JpaRepository<Message, Long> {
+    public interface MessageRepository extends JpaRepository<Message, Integer> {
     }
